Generalize the progress bar implementations
