+++ lib/src/progress_bar.dart
@@ -1,0 +1,205 @@
+import 'package:chewie/src/chewie_progress_colors.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
+import 'package:video_player/video_player.dart';
+
+class ChewieProgressBarConfiguration {
+  const ChewieProgressBarConfiguration({
+    required this.barHeight,
+    required this.handleHeight,
+    required this.drawShadow,
+  });
+
+  final double barHeight;
+  final double handleHeight;
+  final bool drawShadow;
+}
+
+class VideoProgressBar extends StatefulWidget {
+  VideoProgressBar(
+    this.controller, {
+    required this.configuration,
+    ChewieProgressColors? colors,
+    this.onDragEnd,
+    this.onDragStart,
+    this.onDragUpdate,
+    Key? key,
+  })  : colors = colors ?? ChewieProgressColors(),
+        super(key: key);
+
+  final VideoPlayerController controller;
+  final ChewieProgressColors colors;
+  final ChewieProgressBarConfiguration configuration;
+  final Function()? onDragStart;
+  final Function()? onDragEnd;
+  final Function()? onDragUpdate;
+
+  @override
+  _VideoProgressBarState createState() {
+    return _VideoProgressBarState();
+  }
+}
+
+class _VideoProgressBarState extends State<VideoProgressBar> {
+  void listener() {
+    if (!mounted) return;
+    setState(() {});
+  }
+
+  bool _controllerWasPlaying = false;
+
+  VideoPlayerController get controller => widget.controller;
+
+  @override
+  void initState() {
+    super.initState();
+    controller.addListener(listener);
+  }
+
+  @override
+  void deactivate() {
+    controller.removeListener(listener);
+    super.deactivate();
+  }
+
+  void _seekToRelativePosition(Offset globalPosition) {
+    final box = context.findRenderObject() as RenderBox;
+    final Offset tapPos = box.globalToLocal(globalPosition);
+    final double relative = tapPos.dx / box.size.width;
+    final Duration position = controller.value.duration * relative;
+    controller.seekTo(position);
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onHorizontalDragStart: (DragStartDetails details) {
+        if (!controller.value.isInitialized) {
+          return;
+        }
+        _controllerWasPlaying = controller.value.isPlaying;
+        if (_controllerWasPlaying) {
+          controller.pause();
+        }
+
+        widget.onDragStart?.call();
+      },
+      onHorizontalDragUpdate: (DragUpdateDetails details) {
+        if (!controller.value.isInitialized) {
+          return;
+        }
+        _seekToRelativePosition(details.globalPosition);
+
+        widget.onDragUpdate?.call();
+      },
+      onHorizontalDragEnd: (DragEndDetails details) {
+        if (_controllerWasPlaying) {
+          controller.play();
+        }
+
+        widget.onDragEnd?.call();
+      },
+      onTapDown: (TapDownDetails details) {
+        if (!controller.value.isInitialized) {
+          return;
+        }
+        _seekToRelativePosition(details.globalPosition);
+      },
+      child: Center(
+        child: Container(
+          height: MediaQuery.of(context).size.height,
+          width: MediaQuery.of(context).size.width,
+          color: Colors.transparent,
+          child: CustomPaint(
+            painter: _ProgressBarPainter(
+              controller.value,
+              widget.colors,
+              widget.configuration,
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
+
+class _ProgressBarPainter extends CustomPainter {
+  _ProgressBarPainter(this.value, this.colors, this.configuration);
+
+  VideoPlayerValue value;
+  ChewieProgressColors colors;
+
+  ChewieProgressBarConfiguration configuration;
+
+  @override
+  bool shouldRepaint(CustomPainter painter) {
+    return true;
+  }
+
+  @override
+  void paint(Canvas canvas, Size size) {
+    final barHeight = configuration.barHeight;
+    final handleHeight = configuration.handleHeight;
+    final baseOffset = size.height / 2 - barHeight / 2;
+
+    canvas.drawRRect(
+      RRect.fromRectAndRadius(
+        Rect.fromPoints(
+          Offset(0.0, baseOffset),
+          Offset(size.width, baseOffset + barHeight),
+        ),
+        const Radius.circular(4.0),
+      ),
+      colors.backgroundPaint,
+    );
+    if (!value.isInitialized) {
+      return;
+    }
+    final double playedPartPercent =
+        value.position.inMilliseconds / value.duration.inMilliseconds;
+    final double playedPart =
+        playedPartPercent > 1 ? size.width : playedPartPercent * size.width;
+    for (final DurationRange range in value.buffered) {
+      final double start = range.startFraction(value.duration) * size.width;
+      final double end = range.endFraction(value.duration) * size.width;
+      canvas.drawRRect(
+        RRect.fromRectAndRadius(
+          Rect.fromPoints(
+            Offset(start, baseOffset),
+            Offset(end, baseOffset + barHeight),
+          ),
+          const Radius.circular(4.0),
+        ),
+        colors.bufferedPaint,
+      );
+    }
+    canvas.drawRRect(
+      RRect.fromRectAndRadius(
+        Rect.fromPoints(
+          Offset(0.0, baseOffset),
+          Offset(playedPart, baseOffset + barHeight),
+        ),
+        const Radius.circular(4.0),
+      ),
+      colors.playedPaint,
+    );
+
+    if (configuration.drawShadow) {
+      final shadowPath = Path()
+        ..addOval(
+          Rect.fromCircle(
+            center: Offset(playedPart, baseOffset + barHeight / 2),
+            radius: handleHeight,
+          ),
+        );
+
+      canvas.drawShadow(shadowPath, Colors.black, 0.2, false);
+    }
+
+    canvas.drawCircle(
+      Offset(playedPart, baseOffset + barHeight / 2),
+      handleHeight,
+      colors.handlePaint,
+    );
+  }
+}
